为 zako info 添加 操作员和添加时间 字段

+++ src/main/java/com/tty/entity/WhitelistInstance.java
@@ -14,7 +14,7 @@ public class WhitelistInstance {
     @TableField("player_uuid")
     private String playerUUID;
     @TableField("add_time")
-    private double addTime;
+    private long addTime;
     @TableField("operator")
     private String operator;
 }
